+++ workers/ohsome_quality_analyst/oqt.py
@@ -220,6 +220,8 @@ async def _(
     indicator.calculate()
     logging.info("Run figure creation")
     indicator.create_figure()
+    if hasattr(indicator, "create_figure_plotly"):
+        indicator.create_figure_plotly()
     indicator.create_html()
 
     return indicator
@@ -249,6 +251,8 @@ async def _(
     indicator.calculate()
     logging.info("Run figure creation")
     indicator.create_figure()
+    if hasattr(indicator, "create_figure_plotly"):
+        indicator.create_figure_plotly()
     indicator.create_html()
 
     return indicator